Add subtasks structure + Organize css
